requirements.txt
discord.py[voice]
logdna
requests~=2.25.1
youtube-dl
spotipy~=2.16.1
igdb-api-v4
discord-py-slash-command
discord